+++ src/domain/entities/auth.ts
